requirements.txt
Flask==2.3.2
langchain==0.0.263
numpy==1.24.4
openai==0.27.8
python-dotenv==1.0.0
qdrant-client==1.4.0
tiktoken==0.4.0
pypdf==3.15.1
flask-cors==4.0.0
faiss-cpu
pytz
